+++ backend/problems/src/ma1/powers.rs
@@ -1,1 +1,2 @@
+mod non_integer_exponents;
 mod standard_powers;